codeql suggestion fix

+++ tranceiver-lora-sx1278.cpp
@@ -104,7 +104,7 @@ transmit_error_t tranceiver_lora_sx1278::put_message_low(const message & m)
 #endif
 }
 
-tranceiver_lora_sx1278::tranceiver_lora_sx1278(const std::string & id, seen *const s, work_queue_t *const w, gps_connector *const gps, const int dio0_pin, const int reset_pin, const bool digipeater, stats *const st, const int dev_nr) :
+tranceiver_lora_sx1278::tranceiver_lora_sx1278(const std::string & id, seen *const s, work_queue_t *const w, gps_connector *const gps, const int dio0_pin, const int reset_pin, const bool digipeater, stats *const st, const size_t dev_nr) :
 	tranceiver(id, s, w, gps),
 	digipeater(digipeater)
 {
@@ -160,7 +160,7 @@ void tranceiver_lora_sx1278::count_packets(const bool valid)
 		stats_inc_counter(invalid_pkts);
 }
 
-tranceiver *tranceiver_lora_sx1278::instantiate(const libconfig::Setting & node_in, work_queue_t *const w, gps_connector *const gps, stats *const st, const int dev_nr)
+tranceiver *tranceiver_lora_sx1278::instantiate(const libconfig::Setting & node_in, work_queue_t *const w, gps_connector *const gps, stats *const st, const size_t dev_nr)
 {
 	std::string  id;
 	seen        *s          = nullptr;
